Fix : rlwe encryption (#78)
Issue originally discussed in #74 as part of a larger PR.

The io test fails when setting `hardcoded_key_sigma` to any value
different than 0. That's because the bits from the noisy plaintext are
recovered using `extract_highest_bits()`. As an example, if message bit
is 0 and error is negative, the resulting recovered highest bit will be
1 which is not correct

The new logic to recover the message bits implemented in this PR is
based on
[hackmd](https://hackmd.io/TqUeL_F8QMCdAq3dxMbhkg#RLWE-secret-key-encryption)

Furthermore the obfuscate interface is modify to accept the hardcoded
key as input

I also create an independent function `rlwe_encrypt` to perform such
encryption because this is also useful in other parts of the codebase. I
struggled a bit to define the generic types of the inputs to make it
compatible with the `obfuscate` function so please let me know if we can
simplify that.