feat: Introduce ephemeral self-hosted runners for workflows (#131)
With this PR, the two workflows have been migrated to run on ephemeral
self-hosted runners.
The CI workflow runs on a c7i.4xlarge instance and is triggered when
pushing to main.
The Bench workflow runs on an i4i.16xlarge instance and is triggered
manually.

A mechanism was implemented that checks whether the instances can be
launched in the first Availability Zone (AZ). If not, it scans for
subnets in other AZs. That helps with the InsufficientInstanceCapacity
problem.