fix: overflow while conversion (#154)
This is a real fix for PLT where the assertion is not working
: the issue was on `from_usize_to_lsb` method was previously overflowed
at u64, so in real parameter case as the polynomial demension is over 64
bit, this lsb bit representation had repeated instead on filling up in
0. This is different behavior on our plaintext polynomial which is fully
constant polynomial(only first few bytes are filled rest is 0).

Also added some refactorings on methods as it was unclear what format of
polynomial it returns via method name.

fixed benchmark:
https://github.com/MachinaIO/diamond-io/actions/runs/16595490483/job/46941038176